update(app): Transaction wrappers.
Added classes for HardCreate, HardDeposit, SoftTransfer, SoftWithdrawal.
All of them can encode the transactions for the block's transactions data buffer as well as for the transactions tree.
The soft transaction classes have validation functions to ensure that the transactions are able to be executed. These functions receive an account object, validate the nonce, verify the signature, ensure the signer is approved for the account, and check that the balance of the account is sufficient for the transaction value.